feat: use a common api validation
