+++ apps/namadillo/src/atoms/settings/services.ts
@@ -2,17 +2,7 @@ import { DefaultApi } from "@anomaorg/namada-indexer-client";
 import toml from "toml";
 import { SettingsTomlOptions } from "types";
 
-export const isIndexerAlive = async (url: string): Promise<boolean> => {
-  try {
-    const api = new DefaultApi({ basePath: url });
-    const response = await api.healthGet();
-    return response.status === 200;
-  } catch {
-    return false;
-  }
-};
-
-export const isRpcAlive = async (url: string): Promise<boolean> => {
+export const isApiAlive = async (url: string): Promise<boolean> => {
   try {
     const api = new DefaultApi({ basePath: url });
     const response = await api.healthGet();
